support insert into statement

+++ fe/fe-core/src/main/java/org/apache/doris/qe/runtime/BackendFragmentId.java
@@ -24,5 +24,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class BackendFragmentId {
     public final long backendId;
-    public final long fragmentId;
+    public final int fragmentId;
 }
